avocado.core: move methods from jobdata to data_dir
The jobdata module contain helpers to get data from the
jobdata directory in the job results directory. So
the get_resultsdir() and get_id() methods, which get information
from the job results directory - conceptually does not belong to
jobdata. Instead, there is the data_dir module that deal with
the job data directory. Therefore this change move those methods
from jobdata to data_dir.
